back: hoist evaluate() to the Backend base class
