+++ src/xtc/backends/mlir/MlirBackend.py
@@ -5,8 +5,6 @@
 from abc import abstractmethod
 from typing import Any
 from typing_extensions import override
-from pathlib import Path
-import tempfile
 
 from xdsl.dialects import func as xdslfunc
 
@@ -29,8 +27,13 @@ class MlirBackend(itf.back.Backend):
         self.no_alias = no_alias
         self.always_vectorize = always_vectorize
         self.concluding_passes = concluding_passes
-        self.payload_name = str(xdsl_func.sym_name).replace('"', "")
+        self._payload_name = str(xdsl_func.sym_name).replace('"', "")
         self._graph = graph
+
+    @property
+    @override
+    def payload_name(self) -> str:
+        return self._payload_name
 
     @property
     @override
@@ -53,24 +56,3 @@ class MlirBackend(itf.back.Backend):
     @abstractmethod
     def np_outputs_spec(self) -> list[dict[str, Any]]:
         pass
-
-    def evaluate(
-        self,
-        schedule: itf.schd.Schedule,
-        compiler_args: dict = {},
-        evaluate_args: dict = {},
-    ) -> float | str:
-        with tempfile.TemporaryDirectory() as dirname:
-            libpath = Path(dirname) / f"payload_{self.payload_name}"
-            compiler = self.get_compiler(
-                dump_file=str(libpath),
-                shared_lib=True,
-                **compiler_args,
-            )
-            module = compiler.compile(schedule)
-            evaluator = module.get_evaluator(
-                validate=True,
-                **evaluate_args,
-            )
-            results, code, error_msg = evaluator.evaluate()
-        return min(results) if code == 0 else error_msg
